+++ lib/src/models/abstract/text_logger.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:vit_logger/src/models/index.dart';
 import 'package:vit_logger/src/usecases/get_timestamp_string.dart';
 
@@ -42,7 +40,7 @@ abstract class TextLogger extends BaseLogger {
   }
 
   /// Used to indicate informational logs.
-  FutureOr<void> info(String message) async {
+  void info(String message) async {
     await log(
       message: digest(message, LogLevel.info),
       level: LogLevel.info,
@@ -50,7 +48,7 @@ abstract class TextLogger extends BaseLogger {
   }
 
   /// Used to indicate warning logs.
-  FutureOr<void> warn(String message) async {
+  void warn(String message) async {
     await log(
       message: digest(message, LogLevel.warn),
       level: LogLevel.warn,
@@ -58,7 +56,7 @@ abstract class TextLogger extends BaseLogger {
   }
 
   /// Used to indicate error logs.
-  FutureOr<void> error(String message) async {
+  void error(String message) async {
     await log(
       message: digest(message, LogLevel.error),
       level: LogLevel.error,
@@ -66,7 +64,7 @@ abstract class TextLogger extends BaseLogger {
   }
 
   /// Used to indicate debug logs.
-  FutureOr<void> debug(String message) async {
+  void debug(String message) async {
     await log(
       message: digest(message, LogLevel.debug),
       level: LogLevel.debug,