+++ lib/src/models/enums/terminal_printer.dart
@@ -1,0 +1,7 @@
+enum TerminalPrinter {
+  /// Uses `print` method from dart.
+  print,
+
+  /// Uses `log` method from "import 'dart:developer'".
+  developerLog,
+}